requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
et-xmlfile==1.1.0
idna==3.3
numpy==1.22.0
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.5
pyocclient==0.6
PySide2==5.15.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
shiboken2==5.15.2
shiboken6==6.2.2.1
six==1.16.0
urllib3==1.26.8

QDarkStyle~=3.0.3